Move epa_check_alignment from bsg_manycore.cpp into bsg_manycore_epa.h
